Organized the code into managable dirs
